@@ -1,16 +1,19 @@
 {
   "name": "StarterKit",
-  "version": "0.0.1",
+  "version": "0.10.0",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "^0.22.2",
+    "react": "^15.0.2",
+    "react-native": "^0.26.1",
     "react-native-db-models": "^0.1.3",
-    "react-native-navbar": "^1.4.2",
-    "react-native-side-menu": "^0.18.0",
-    "tcomb-form-native": "^0.3.3"
+    "react-native-navbar": "^1.5.0",
+    "react-native-side-menu": "^0.19.0",
+    "tcomb-form-native": "^0.5.0"
+  },
+  "devDependencies": {
+    "rnpm-plugin-upgrade": "^0.26.0"
   }
 }
